shift RuleForm rendering out of RuleList
